+++ src/richie/apps/courses/settings/__init__.py
@@ -597,3 +597,14 @@ RICHIE_SIMPLEPICTURE_PRESETS = {
         "sizes": "60px",
     },
 }
+
+# If true the toolbar item will already be showed. If false only a page which already
+# have the extension will have the toolbar item and users won't be able to add
+# MainMenuEntry extension on existing page, only create new page with index extension
+# through the wizard.
+RICHIE_MAINMENUENTRY_ALLOW_CREATION = False
+
+# Define which node level can be processed to search for MainMenuEntry extension. You
+# can set it to 'None' for never processing any node.
+# This is a limit against performance issues to avoid making querysets for nothing.
+RICHIE_MAINMENUENTRY_MENU_ALLOWED_LEVEL = 0